Refactor imports to utils and fix token getter
Move shared helpers and constants into utils: apiMeteo now imports encodeQueryData from '@/utils/helpers' and constants.ts uses HOST_IMG from '@/utils/constants'. Simplify getStorageToken to avoid duplicate LocalStorage lookups and handle SSR by returning an empty string when no token is available. Update client/api/index.ts to re-export HOST_API, HOST_IMG, SITE_LINK and encodeQueryData from the utils modules and tidy top-level exports.
